api/requirements.txt
fastapi==0.115.4
sqlmodel==0.0.22
asyncpg==0.30.0
uvicorn==0.32.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10